move hook to sep file
